third_party/WebKit/LayoutTests/media/video-preload-datasaver.html

<!DOCTYPE html>
<html>
<head>
<title>Test preload behaviour with data saver setting</title>
</head>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-file.js"></script>
<script>
  window.internals.settings.setDataSaverEnabled(true);

  async_test(function(t) {
    var v = document.createElement('video');
    v.src = findMediaFile('video', 'content/test');

    assert_equals(v.preload, 'none');
    v.onsuspend = t.step_func_done();
    v.onprogress = t.unreached_func();
  }, "Data saver makes the default preload value to 'none'.");

  async_test(function(t) {
    var v = document.createElement('video');
    v.src = findMediaFile('video', 'content/test');
    v.preload = 'metadata';

    assert_equals(v.preload, 'none');
    v.onsuspend = t.step_func_done();
    v.onprogress = t.unreached_func();
  }, "Data saver overrides preload='metadata'.");

  async_test(function(t) {
    var v = document.createElement('video');
    v.src = findMediaFile('video', 'content/test');
    v.preload = 'auto';

    assert_equals(v.preload, 'none');
    v.onsuspend = t.step_func_done();
    v.onprogress = t.unreached_func();
  }, "Data saver overrides preload='auto'.");

  async_test(function(t) {
    window.internals.settings.setMediaPlaybackRequiresUserGesture(true);

    var v = document.createElement('video');
    v.autoplay = true;
    v.src = findMediaFile('video', 'content/test');

    assert_equals(v.preload, 'none');
    v.onsuspend = t.step_func_done();
    v.onprogress = t.unreached_func();

    t.add_cleanup(function() {
      window.internals.settings.setMediaPlaybackRequiresUserGesture(false);
    });
  }, "Data saver prevents loading autoplaying videos when autoplay is blocked.");
</script>
</body>
</html>
